+++ src/routes/generalRoutes.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const generalController = require('../controllers/generalController');
+
+router.get('/orders', generalController.getOrders);
+router.get('/budget', generalController.getBudget);
+module.exports = router;
