--- src/routes/orderRoutes.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const orderController = require('../controllers/orderController');
-
-router.get('/', orderController.getOrders);
-
-module.exports = router;
