+++ src/server.js
@@ -4,7 +4,7 @@ const cookieParser = require('cookie-parser');
 require('dotenv').config();
 
 const authRoutes = require('./routes/authRoutes');
-const orderRoutes = require('./routes/orderRoutes');
+const generalRoutes = require('./routes/generalRoutes');
 const therapistRoutes = require('./routes/therapistRoutes');
 
 const app = express();
@@ -46,7 +46,7 @@ app.use((req, res, next) => {
 });
 
 app.use('/auth', authRoutes);
-app.use('/orders', orderRoutes);
+app.use('/', generalRoutes);
 app.use('/therapist', therapistRoutes);
